Move date of birth information to personal step
- Move DOB info to personal step from ID step
